+++ packages/dev/devServer/testsuite/all/devSite throw 2.ja.ts
@@ -1,8 +1,6 @@
-import { getRandomInteger, sleeping } from "^jab";
+import { sleeping } from "^jab";
 import { TestProvision } from "^jarun";
 
 export default (prov: TestProvision) => {
-  console.log(getRandomInteger(234));
-
   return sleeping(500);
 };