+++ packages/tests/all/jatev createSetTestCaseUpdater err 2.ja.ts
@@ -3,7 +3,9 @@ import { TestProvision } from "^jarun";
 
 import { getSetTestCaseUpdate_with_tests } from "../_fixture";
 
-// test id doesn't exist
+// test id doesn't exist is just ignored
+
+//flexibility needed when changed test selections.
 
 export default (prov: TestProvision) => {
   prov.imp(